add : 표준 응답 형태 정의

+++ src/main/java/com/jejeong/apipractice/controller/sign/SignController.java
@@ -7,6 +7,7 @@ import com.jejeong.apipractice.sevice.member.SignService;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 @RestController
@@ -18,20 +19,26 @@ public class SignController {
 
     @PostMapping("/sign-up")
     @ResponseStatus(HttpStatus.CREATED)
-    public void signUpRequest(@Valid @RequestBody SignUpRequest req) {
+    public ResponseEntity<Valid> signUpRequest(@Valid @RequestBody SignUpRequest req) {
         memberService.signUp(req);
+        return ResponseEntity.status(HttpStatus.CREATED).build();
     }
 
     @GetMapping("/me")
-    public SignResponse processUserInfoRequest(String email) {
+    public ResponseEntity<SignResponse> processUserInfoRequest(@RequestParam String email) {
         // TODO email 대신 사용자 정보로 조회할 수 있게끔 해야함.
         MemberDto member = memberService.loadUserByUserEmail(email);
-        return SignResponse.of(member.getEmail(), member.getNickname());
+        if (member != null)
+            return ResponseEntity.ok(SignResponse.of(member.getEmail(), member.getNickname()));
+        return ResponseEntity.notFound().build();
     }
 
     @DeleteMapping("/me")
-    public void processWithdrawalRequest(String email) {
+    public ResponseEntity<Void> processWithdrawalRequest(@RequestParam String email) {
         // TODO email 대신 사용자 정보로 조회할 수 있게끔 해야함.
+        if (email.isBlank()) return ResponseEntity.badRequest().build();
+
         memberService.deleteUserByUserEmail(email);
+        return ResponseEntity.status(HttpStatus.OK).build();
     }
 }
